* Navbar was moved to Organism which is more truthy with files structure since Navbar includes more molecules
