+++ src/App.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Route, Routes } from "react-router-dom";
 import { LoginPage } from "./Pages/LoginPage/LoginPage";
 import { CvPage } from "./Pages/CvPage";
+import { AvailableStudentsPage } from "./Pages/AvailableStudentsPage/AvailableStudentsPage";
 
 function App() {
   return (
@@ -13,6 +14,10 @@ function App() {
       <Route
         path="/cv"
         element={<CvPage />}
+      />
+      <Route
+        path="/available-students"
+        element={<AvailableStudentsPage />}
       />
     </Routes>
   );
