+++ src/Pages/CvPage.tsx
@@ -1,5 +1,12 @@
+import React from "react";
 import { CvOrganism } from "../components/Organism/CvOrganism";
+import { Navbar } from "../components/Organism/Navbar";
 
 export function CvPage() {
-  return <CvOrganism />;
+  return (
+    <>
+      <Navbar />
+      <CvOrganism />
+    </>
+  );
 }
