+++ src/components/Atoms/NavModalContainer.tsx
@@ -16,7 +16,7 @@ export const NavModalContainer = styled.div`
   position: fixed;
   z-index: 2;
   animation: showMe 0.3s linear forwards;
-
+  box-shadow: 0 0 5px 5px rgba(0, 0, 0, 0.3);
   @keyframes showMe {
     0% {
       transform: translateY(-100%);
